refactor: centralize design tokens in @drafto/shared
Move palette, semantic light/dark, spacing, radii, and fontSizes tokens
into packages/shared/src/design-tokens.ts as the single source of truth.
Mobile and desktop theme/tokens.ts become re-export shims so the
@/theme/tokens import surface is unchanged.

Add a drift test in packages/shared that parses apps/web/src/app/globals.css
and asserts every palette and semantic CSS custom property matches the TS
token value. The test already flagged one pre-existing drift: the
dark-mode --ring variable was left at the old palette value (#a5b4fc)
when the Digital Atelier palette landed in ADR 0014 — reconciled by
updating globals.css to match the TS source of truth (#9478ff =
colors.primary[300]).

Also add ADR 0020 documenting the decision.
