+++ apps/web/src/app/globals.css
@@ -54,7 +54,7 @@
   --fg-subtle: #6b6360;
   --border: rgba(199, 196, 216, 0.1);
   --border-strong: rgba(199, 196, 216, 0.2);
-  --ring: #a5b4fc;
+  --ring: #9478ff;
   --sidebar-bg: #251f1b;
   --sidebar-hover: #2e2822;
   --sidebar-active: #2a2250;
