+++ packages/shared/package.json
@@ -13,6 +13,7 @@
     "test": "vitest run"
   },
   "devDependencies": {
+    "@types/node": "^24",
     "typescript": "^6",
     "vitest": "^4.1.2"
   }
