+++ packages/shared/tsconfig.json
@@ -10,7 +10,8 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "incremental": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "types": ["node"]
   },
   "include": ["src/**/*.ts", "__tests__/**/*.ts"],
   "exclude": ["node_modules"]
